requirements.txt
Flask==2.0.1
Flask-RESTful
google-cloud-firestore
gunicorn
Werkzeug